Add a real schema
- Factor our dump method into translation.utils (may be the wrong place)
- Add schema autogenerated from normalized pexel example @autoencoder provided
- Pass translation tests
