+++ mediachain/reader/api.py
@@ -4,24 +4,11 @@ import copy
 import base58
 import json
 from pygments import highlight
-from pygments.lexers.data import JsonLexer
-from pygments.formatters.terminal import TerminalFormatter
+from mediachain.translation.utils import dump
 
 def get_and_print_object(transactor, object_id):
     obj = get_object(transactor, object_id, fetch_images=False)
-    j = json.dumps(stringify_refs(obj), indent=2)
-    print highlight(j, JsonLexer(), TerminalFormatter())
-
-
-def stringify_refs(obj):
-    res = {}
-    for k, v in obj.iteritems():
-        if isinstance(v, dict):
-            v = stringify_refs(v)
-        if k == u'@link':
-            v = base58.b58encode(v)
-        res[k] = v
-    return res
+    dump(obj)
 
 
 def get_object(transactor, object_id, fetch_images=True):
